requirements.txt
python-telegram-bot==21.3
sqlalchemy==2.0.35
aiosqlite==0.20.0
aiohttp==3.10.5
python-dotenv==1.0.1
greenlet==3.0.3
